// src/modules/auth/enums/permission.enum.ts
export enum PermissionCodeEnum {
  READ_USER = 'Melihat Pengguna',
  CREATE_USER = 'Membuat Pengguna',
  UPDATE_USER = 'Mengubah Pengguna',
  DELETE_USER = 'Menghapus Pengguna',

  READ_ROLE = 'Melihat Peran',
  CREATE_ROLE = 'Membuat Peran',
  UPDATE_ROLE = 'Mengubah Peran',
  DELETE_ROLE = 'Menghapus Peran',
}

export enum PermissionGroupEnum {
  AUTH = 'Autentikasi',
}